project_1/requirements.txt
glfw==2.7.0
numpy==1.26.4
pyopengl==3.1.7
tabulate==0.9.0
